[refactor] : DurationMap 관련 파일 Duration 디렉토리로 이동
