+++ src/main/java/com/halo/eventer/domain/duration/repository/DurationMapRepository.java
@@ -1,6 +1,8 @@
-package com.halo.eventer.domain.duration_map;
+package com.halo.eventer.domain.duration.repository;
 
 import java.util.List;
+
+import com.halo.eventer.domain.duration.DurationMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DurationMapRepository extends JpaRepository<DurationMap, Long> {
